T-13: los tres bloqueantes del verificador — medir donde antes había una afirmación
Los tres eran el mismo error en tres sitios: una frase bien escrita ocupando el
sitio de una medida. Y el método para cerrarlos ha sido el que faltaba: medir
sobre lo que se publica y comprobar que la sonda llega al artefacto.

B1 · El golden del coeficiente se ensanchó sin necesidad y con un argumento
falso. «Quitar MA2/MB2 no lo arregla» sólo era cierto para la cota de sesgo, que
es la aserción secundaria, y se usó para justificar el cambio de la principal.
Revertido: `toleranceUnits` vuelve a 2. Lo que cambia es el instrumento —MA2 y
MB2 son modulación radiacional, no marea astronómica, y salen del cálculo del
coeficiente, no del dataset—, con lo que los 32 valores publicados vuelven a
caer dentro de ±2. Sólo se ensancha la cota del sesgo agregado, de 1 a 1,25,
midiendo +1,19, y se explica por qué no se saca además EP2 aunque dejaría el
sesgo en 0,84: es marea astronómica pura y elegirlo por lo bien que le sienta al
golden es el mismo pecado con otro traje. Y se corrige el pilar: el coste de
truncado no se mide contra el IOC —es el modelo contra sí mismo y baja por
definición—, lo que sí mira al mar empeora un pelo (RMSE 0,0794 → 0,0806) y el
grade A de Brest es mecánico.

B2 · El atajo de T-05 se había retirado en el dato pero sin trinquete: inyectar
a mano el RMSE de un mareógrafo a 46 km dejaba la suite en verde. Ahora el
dataset publica la procedencia del número —con qué mareógrafo se midió y a qué
distancia de la dársena— y hay un invariante en Python y en TypeScript que exige
que todo RMSE publicado se haya medido a menos de 5 km. Reproducida la inyección
en tres formas, las tres salen en rojo. Con las cuatro ramas de estimate() y un
test que fija que la observación se busca una sola vez y en el puerto.

B3 · El aviso afirmaba para todo puerto estimado que «no publicamos el error, no
lo hemos medido», y Garachico y San Sebastián de la Gomera lo publican veinte
líneas más abajo en la misma página. Es el reverso del pecado de T-05: negar un
número propio. El aviso tiene ahora dos versiones y un test que compara lo que
afirma con lo que la página publica; en Garachico esos 126 min negados eran la
mejor prueba de que sus constantes prestadas no valen ahí.

Y los dos no bloqueantes, que eran de la misma familia: el gate A-1 se re-apunta
sin ningún número inventado (la meseta es legítima si la marea se movió menos
que el paso de publicación, 1 mm; medido 0,983 mm máximo sobre 500 mesetas), el
motivo del grade deja de callar la mitad en 39 puertos, y CI empieza a lintear
Python con ruff pinneado — que en su primer pase encontró una variable muerta.

+++ data/pipeline/mareia_pipeline/tides/predict.py
@@ -59,7 +59,7 @@ def nodal_factors(harmonics: list[Harmonic], centre: dt.datetime) -> dict[str, t
 
 def time_grid(start: dt.datetime, days: float, step_minutes: float) -> np.ndarray:
     """Rejilla de instantes UTC como horas desde la época de Schureman."""
-    count = int(round(days * 24 * 60 / step_minutes)) + 1
+    count = round(days * 24 * 60 / step_minutes) + 1
     base = hours_since_epoch(start)
     return base + np.arange(count) * (step_minutes / 60.0)
